requirements.txt
pandas>=2.0.0
pyyaml>=6.0
snowflake-connector-python>=3.0.0
pytest>=7.0.0
